# buscatuprofe-back/docs/swagger.yaml
definitions:
  models.DTOAddClase:
    properties:
      descripcion:
        type: string
      duracion:
        type: string
      materias:
        items:
          type: string
        type: array
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        description: ID          uint    `json:"id"`
        type: string
      precio:
        type: number
      profesor_id:
        type: integer
      provincia_id:
        type: integer
    type: object
  models.DTOAllClase:
    properties:
      descripcion:
        type: string
      duracion:
        type: string
      id:
        type: integer
      materias:
        items:
          type: string
        type: array
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      precio:
        type: number
      profesor:
        type: string
      provincia:
        type: string
    type: object
  models.DTOClaseID:
    properties:
      descripcion:
        type: string
      duracion:
        type: string
      id:
        type: integer
      materias:
        items:
          type: string
        type: array
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      precio:
        type: number
      profesor:
        $ref: '#/definitions/models.DTOProfesor'
      provincia:
        type: string
    type: object
  models.DTOClasesProfesor:
    properties:
      descripcion:
        type: string
      duracion:
        type: string
      id:
        type: integer
      materias:
        items:
          type: string
        type: array
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      precio:
        type: number
      provincia:
        type: string
    type: object
  models.DTOMateria:
    properties:
      materias:
        items:
          type: string
        type: array
    type: object
  models.DTOProfesor:
    properties:
      disponibilidad:
        items:
          type: string
        type: array
      id:
        type: integer
      mail:
        type: string
      nombre:
        type: string
      telefono:
        type: string
      titulos:
        items:
          type: string
        type: array
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: no se pudieron obtener las clases
        type: string
    type: object
info:
  contact: {}
paths:
  /clases:
    get:
      description: Retorna la informacion de cada clase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DTOAllClase'
            type: array
        "500":
          description: No se pudieron obtener las clases
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Obtiene todas las clases
      tags:
      - Clases
    post:
      consumes:
      - application/json
      description: Agrega una nueva clase
      parameters:
      - description: Datos de la clase a crear
        in: body
        name: clase
        required: true
        schema:
          $ref: '#/definitions/models.DTOAddClase'
      produces:
      - application/json
      responses:
        "200":
          description: Clase guardada correctamente
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Datos invalidos o faltantes
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Error interno al guardar la clase
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Crear una clase
      tags:
      - Clases
  /clases/{id}:
    get:
      description: Retorna la informacion de una clase especifica
      parameters:
      - description: ID de la clase
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DTOClaseID'
        "400":
          description: Entrada inválida
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Clase no encontrada
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Error interno
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Obtiene una clase por id
      tags:
      - Clases
  /materias:
    get:
      description: Obtiene los datos de las materias disponibles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DTOMateria'
        "500":
          description: No se pudieron obtener las materias
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Lista todas las materias disponibles
      tags:
      - Materias
  /profesor/clases/{id}:
    get:
      description: Obtiene los datos de las clases de un profesor
      parameters:
      - description: ID del profesor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DTOClasesProfesor'
            type: array
        "500":
          description: No se pudieron obtener las clases
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Obtener las clases de un profesor
      tags:
      - Clases
swagger: "2.0"